cleanup old apis, KBarResults by default virtualized
