+++ src/KBarResults.tsx
@@ -1,232 +1,168 @@
-import { matchSorter } from "match-sorter";
 import * as React from "react";
-import { VisualState } from "./types";
-import type {
-  Action,
-  KBarResultsProps,
-  ResultHandlers,
-  ResultState,
-} from "./types";
-import useKBar from "./useKBar";
+import { useVirtual } from "react-virtual";
+import { useKBar } from ".";
+import { Action } from "./types";
 
-/**
- * @deprecated Results can be created using a combination of the `Results` component alongside
- * `useResultItem` hook.
- *
- * ```tsx
- * import { Results, useResultItem } from "kbar";
- * ```
- *
- * Reference: https://github.com/timc1/kbar/pull/59
- */
-export default function KBarResults(props: KBarResultsProps) {
-  const { search, actions, currentRootActionId, query } = useKBar((state) => ({
+const START_INDEX = 0;
+
+interface RenderParams<T = Action | string> {
+  item: T;
+  active: boolean;
+}
+
+interface KBarResultsProps {
+  items: any[];
+  onRender: (params: RenderParams) => React.ReactElement;
+  maxHeight?: number;
+}
+
+const KBarResults: React.FC<KBarResultsProps> = (props) => {
+  const activeRef = React.useRef<HTMLDivElement>(null);
+  const parentRef = React.useRef(null);
+
+  // store a ref to all items so we do not have to pass
+  // them as a dependency when setting up event listeners.
+  const itemsRef = React.useRef(props.items);
+  itemsRef.current = props.items;
+
+  const rowVirtualizer = useVirtual({
+    size: itemsRef.current.length,
+    parentRef,
+  });
+
+  const [activeIndex, setActiveIndex] = React.useState(START_INDEX);
+
+  const { query, search, currentRootActionId } = useKBar((state) => ({
     search: state.searchQuery,
     currentRootActionId: state.currentRootActionId,
-    actions: state.actions,
   }));
 
-  // Store reference to a list of all actions
-  const actionsList = React.useMemo(
-    () =>
-      Object.keys(actions).map((key) => {
-        return actions[key];
-      }),
-    [actions]
-  );
-
-  const currActions = React.useMemo(() => {
-    if (
-      !currentRootActionId ||
-      // `currentRootActionId` can refer to an action that is no
-      // longer registered while kbar is open in the case where
-      // `useRegisterActions`'s effect is ran multiple times.
-      (currentRootActionId && !actions[currentRootActionId])
-    ) {
-      return actionsList.reduce((acc, curr) => {
-        if (!curr.parent) {
-          acc[curr.id] = curr;
-        }
-        return acc;
-      }, {});
-    }
-
-    const root = actions[currentRootActionId];
-    const children = root.children;
-
-    if (!children) {
-      return {
-        [root.id]: root,
-      };
-    }
-
-    return {
-      ...children.reduce((acc, actionId) => {
-        acc[actionId] = actions[actionId];
-        return acc;
-      }, {}),
-    };
-  }, [actions, actionsList, currentRootActionId]);
-
-  const filteredList = React.useMemo(
-    () =>
-      Object.keys(currActions).map((key) => {
-        const action = currActions[key];
-        return action;
-      }) as Action[],
-    [currActions]
-  );
-
-  const matches = useMatches(search, filteredList);
-
-  const [activeIndex, setActiveIndex] = React.useState(0);
-
-  // Reset active index on root action change
   React.useEffect(() => {
-    setActiveIndex(0);
-  }, [currentRootActionId]);
-
-  const select = React.useCallback(() => {
-    const action = matches[activeIndex];
-
-    if (!action) return;
-
-    if (action.perform) {
-      action.perform();
-      query.setVisualState(() => VisualState.animatingOut);
-    } else {
-      query.setCurrentRootAction(action.id);
-    }
-  }, [matches, activeIndex, query]);
-
-  React.useEffect(() => {
-    function handleKeyDown(event: KeyboardEvent) {
-      event.stopPropagation();
-
-      if (event.key === "ArrowDown" || (event.ctrlKey && event.key === "n")) {
-        event.preventDefault();
-        setActiveIndex((index) => {
-          if (index >= matches.length - 1) {
-            return 0;
-          } else {
-            return index + 1;
-          }
-        });
-      }
-
+    const handler = (event) => {
       if (event.key === "ArrowUp" || (event.ctrlKey && event.key === "p")) {
         event.preventDefault();
         setActiveIndex((index) => {
-          if (index === 0) {
-            return matches.length - 1;
-          } else {
-            return index - 1;
+          let nextIndex = index > START_INDEX ? index - 1 : index;
+          // avoid setting active index on a group
+          if (typeof itemsRef.current[nextIndex] === "string") {
+            nextIndex -= 1;
           }
+          return nextIndex;
         });
-      }
-
-      if (event.key === "Enter") {
+      } else if (
+        event.key === "ArrowDown" ||
+        (event.ctrlKey && event.key === "n")
+      ) {
         event.preventDefault();
-        select();
-      }
-    }
-
-    window.addEventListener("keydown", handleKeyDown);
-    return () => window.removeEventListener("keydown", handleKeyDown);
-  }, [matches, select, activeIndex]);
-
-  // Reset focused index when searching & updated results.
-  React.useEffect(() => {
-    setActiveIndex(0);
-  }, [filteredList.length, search]);
-
-  return (
-    <div className={props.className} style={props.style}>
-      {matches.length
-        ? matches.map((action, index) => {
-            const handlers: ResultHandlers = {
-              onClick: select,
-              onPointerDown: () => setActiveIndex(index),
-              onMouseEnter: () => setActiveIndex(index),
-            };
-
-            const state: ResultState = {
-              activeIndex,
-              index,
-            };
-
-            if (props.onRender) {
-              // Implicitly add a `key` so the user won't have to.
-              return React.cloneElement(
-                props.onRender(action, handlers, state),
-                {
-                  key: action.id,
-                }
-              );
-            }
-
-            return (
-              <DefaultResultWrapper
-                key={action.id}
-                isActive={activeIndex === index}
-                {...handlers}
-              >
-                {action.name}
-              </DefaultResultWrapper>
-            );
-          })
-        : null}
-    </div>
-  );
-}
-
-// Separate component to ensure we can scrollTo active elements properly.
-const DefaultResultWrapper: React.FC<{ isActive: boolean }> = ({
-  isActive,
-  children,
-  ...rest
-}) => {
-  const ownRef = React.useRef<HTMLDivElement>(null);
-
-  React.useEffect(() => {
-    if (isActive) {
-      // wait for the KBarAnimator to resize, _then_ scrollIntoView.
-      // https://medium.com/@owencm/one-weird-trick-to-performant-touch-response-animations-with-react-9fe4a0838116
-      window.requestAnimationFrame(() =>
-        window.requestAnimationFrame(() => {
-          const element = ownRef.current;
-          if (!element) {
-            return;
+        setActiveIndex((index) => {
+          let nextIndex =
+            index < itemsRef.current.length - 1 ? index + 1 : index;
+          // avoid setting active index on a group
+          if (typeof itemsRef.current[nextIndex] === "string") {
+            nextIndex += 1;
           }
-          element.scrollIntoView({ block: "nearest" });
-        })
-      );
-    }
-  }, [isActive]);
+          return nextIndex;
+        });
+      } else if (event.key === "Enter") {
+        event.preventDefault();
+        // storing the active dom element in a ref prevents us from
+        // having to calculate the current action to perform based
+        // on the `activeIndex`, which we would have needed to add
+        // as part of the dependencies array.
+        activeRef.current?.click();
+      }
+    };
+    window.addEventListener("keydown", handler);
+    return () => window.removeEventListener("keydown", handler);
+  }, []);
+
+  // destructuring here to prevent linter warning to pass
+  // entire rowVirtualizer in the dependencies array.
+  const { scrollToIndex } = rowVirtualizer;
+  React.useEffect(() => {
+    scrollToIndex(activeIndex);
+  }, [activeIndex, scrollToIndex]);
+
+  React.useEffect(() => {
+    // TODO(tim): fix scenario where async actions load in
+    // and active index is reset to the first item. i.e. when
+    // users register actions and bust the `useRegisterActions`
+    // cache, we won't want to reset their active index as they
+    // are navigating the list.
+    setActiveIndex(
+      // avoid setting active index on a group
+      typeof props.items[START_INDEX] === "string"
+        ? START_INDEX + 1
+        : START_INDEX
+    );
+  }, [search, currentRootActionId, props.items]);
+
+  const execute = React.useCallback(
+    (item: RenderParams["item"]) => {
+      if (typeof item === "string") return;
+      if (item.perform) {
+        item.perform();
+        query.toggle();
+        return;
+      }
+      query.setSearch("");
+      query.setCurrentRootAction(item.id);
+    },
+    [query]
+  );
 
   return (
     <div
-      ref={ownRef}
+      ref={parentRef}
       style={{
-        background: isActive ? "#eee" : "#fff",
-        padding: "8px",
+        maxHeight: props.maxHeight || 400,
+        overflow: "auto",
       }}
-      {...rest}
     >
-      {children}
+      <div
+        style={{
+          height: `${rowVirtualizer.totalSize}px`,
+          width: "100%",
+          position: "relative",
+        }}
+      >
+        {rowVirtualizer.virtualItems.map((virtualRow) => {
+          const item = itemsRef.current[virtualRow.index];
+          const handlers = {
+            onMouseEnter: () => setActiveIndex(virtualRow.index),
+            onPointerDown: () => setActiveIndex(virtualRow.index),
+            onClick: () => execute(item),
+          };
+          const active = virtualRow.index === activeIndex;
+
+          return (
+            <div
+              ref={active ? activeRef : null}
+              key={virtualRow.index}
+              style={{
+                position: "absolute",
+                top: 0,
+                left: 0,
+                width: "100%",
+                transform: `translateY(${virtualRow.start}px)`,
+              }}
+              {...handlers}
+            >
+              {React.cloneElement(
+                props.onRender({
+                  item,
+                  active,
+                }),
+                {
+                  ref: virtualRow.measureRef,
+                }
+              )}
+            </div>
+          );
+        })}
+      </div>
     </div>
   );
 };
 
-function useMatches(term: string, actions: Action[]) {
-  // TODO: we can throttle this if needed
-  return React.useMemo(
-    () =>
-      term.trim() === ""
-        ? actions
-        : matchSorter(actions, term, {
-            keys: ["keywords", "name", "subtitle"],
-          }),
-    [term, actions]
-  );
-}
+export default KBarResults;
