--- src/Results/index.tsx
@@ -1,4 +1,0 @@
-import Results from "./Results";
-import { useResultItem } from "./useResultItem";
-
-export { Results, useResultItem };
