+++ src/types.ts
@@ -66,24 +66,3 @@ export enum VisualState {
   animatingOut = "animating-out",
   hidden = "hidden",
 }
-
-export interface ResultHandlers {
-  onClick: () => void;
-  onMouseEnter: () => void;
-  onPointerDown: () => void;
-}
-
-export interface ResultState {
-  index: number;
-  activeIndex: number;
-}
-
-export interface KBarResultsProps {
-  style?: React.CSSProperties;
-  className?: string;
-  onRender?: (
-    action: Action,
-    handlers: ResultHandlers,
-    state: ResultState
-  ) => React.ReactElement;
-}
